Make it compile on both scala versions.

+++ s3/src/test/scala/io/atlassian/aws/s3/S3Spec.scala
@@ -268,7 +268,7 @@ class S3Spec(arguments: Arguments) extends SpecificationWithJUnit with ScalaChec
       val dataStream = new ByteArrayInputStream(data.data)
       val key = S3Key(s"$TEST_FOLDER/${data.key}")
       val location = ContentLocation(BUCKET, key)
-      val invalidRange = Range.From(data.data.length + 1)
+      val invalidRange = Range.From(data.data.length.toLong + 1)
       (for {
         _ <- S3.putStream(location, dataStream, Some(data.data.length.toLong))
         result <- S3.safeGet(location, invalidRange)
